backend/tama_backend/requirements.txt
Django>=4.0,<6.0
python-dateutil
psycopg2-binary
django-redis
daphne
channels
channels-redis
python-decouple
djangorestframework
djangorestframework-simplejwt
django-cors-headers
google-generativeai
regex
pytz
